% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_walk.R
\name{group_walk}
\alias{group_walk}
\title{Implement group-walk algorithm}
\usage{
group_walk(
  winning_scores,
  labels,
  all_group_ids,
  K = 40,
  return_frontier = FALSE,
  correction = 1
)
}
\arguments{
\item{winning_scores}{A numerical vector of winning scores generated from the target-decoy competitions for each hypothesis.}

\item{labels}{A vector of winning labels indicating whether it was a target (= 1) or a decoy (!= 1) for each hypothesis.}

\item{all_group_ids}{A vector of group IDs associated to each hypothesis (can be recorded as integers, factors, characters).}

\item{K}{A window size parameter (integer).}

\item{return_frontier}{A boolean indicating whether the function should return the complete sequence of frontiers.}

\item{correction}{A correction factor used to in the numerator of the estimated false discovery rate (FDR) (Use 1 for FDR control).}
}
\value{
A sequence of q-values for each hypothesis. Users should report the target-hypotheses with q-values less than or equal to their choice of threshold, alpha. If return_frontier = T, additionally the sequence of frontiers will be returned.
}
\description{
Implement group-walk algorithm
}
